isobmff: Cap video/x-raw pad template dimensions to u16::MAX
The isobmff VisualSampleEntry (ISO 14496-12) stores width and height as
unsigned int(16).  Previously isomp4mux and isofmp4mux used u32.MAX which
mean isomp4mux and isofmp4mux would accept resolution higher then the spec.
This change limit resolution accepted to be within the spec. Also fixes
video/x-bayer pad template in isomp4mux which was using i32.MAX while
isofmp4mux was using u16.MAX.

Part-of: <https://gitlab.freedesktop.org/gstreamer/gst-plugins-rs/-/merge_requests/2942>

+++ mux/isobmff/src/isobmff/fmp4mux/imp.rs
@@ -5296,8 +5296,8 @@ impl ElementImpl for ISOFMP4Mux {
                                 "r210",
                             ]),
                         )
-                        .field("width", gst::IntRange::new(1, i32::MAX))
-                        .field("height", gst::IntRange::new(1, i32::MAX))
+                        .field("width", gst::IntRange::new(1, u16::MAX as i32))
+                        .field("height", gst::IntRange::new(1, u16::MAX as i32))
                         .build(),
                     gst::Structure::builder("video/x-raw")
                         // TODO: this could be extended to handle gst_video::VideoMeta for non-default stride and plane offsets
@@ -5308,9 +5308,9 @@ impl ElementImpl for ISOFMP4Mux {
                         )
                         .field(
                             "width",
-                            gst::IntRange::with_step(4, i32::MAX.prev_multiple_of(&4), 4),
+                            gst::IntRange::with_step(4, (u16::MAX as i32).prev_multiple_of(&4), 4),
                         )
-                        .field("height", gst::IntRange::new(1, i32::MAX))
+                        .field("height", gst::IntRange::new(1, u16::MAX as i32))
                         .build(),
                     gst::Structure::builder("video/x-raw")
                         // TODO: this could be extended to handle gst_video::VideoMeta for non-default stride and plane offsets
@@ -5321,11 +5321,11 @@ impl ElementImpl for ISOFMP4Mux {
                         )
                         .field(
                             "width",
-                            gst::IntRange::with_step(4, i32::MAX.prev_multiple_of(&4), 4),
+                            gst::IntRange::with_step(4, (u16::MAX as i32).prev_multiple_of(&4), 4),
                         )
                         .field(
                             "height",
-                            gst::IntRange::with_step(2, i32::MAX.prev_multiple_of(&2), 2),
+                            gst::IntRange::with_step(2, (u16::MAX as i32).prev_multiple_of(&2), 2),
                         )
                         .build(),
                     gst::Structure::builder("video/x-bayer")
